Yet another completions fix: correctly handle commented-out commands
The previous version had a nasty edge case: if there is no comment sign before
the \bibliography{} command, the regexp matched the *entire* text up to and
including that command!
The new version collects all bibliography matches, then uses a for loop toe
iterate until it finds one that is not commented out.