Refs and cites: TAB replaces ref_... and cite_... with R and C respectively
Before showing the drop-down menu, the ref_... and cite_... text gets
replaced with a single letter when the user hits the TAB key.
This leaves more room in the menu for the title of the citation or reference.
The only disadvantage is that, if the user decides against inserting a
citation, the R or C stay in the text---although they can be undone with
CTRL/CMD-Z.