Jump to PDF after compiling. See extended commit notes for details
Currently, there are two limitations:
On both OSX and Windows, multi-file documents are only partially supported.
Compilation can be invoked only from the master document, in which case the
viewer starts.
On Windows, the viewer must already be running (with the correct document in
the front-most window). This will be fixed in a later commit.

In general, multi-file doc support is in its infancy. Log parsing does work,
and next_error does work as a consequence. However, future commits will bring
broader support---in particular, compiling from any file in the project.