Allows changing the format of the citation listings in the auto-complete dialogs.
Added configurable functionality for changing the format of the citation listings in the
auto-complete dialogs (both the small Ctrl-Space dialog and the larger dialog created by
the LatexCiteCommand.

The current implementation works exactly the same as the previous code, as all of the default
configuration settings are set to produce the same results as before. However, it allows easy
configuration to change the display of citation listings. For example, setting

latex_cite_completion_display_line1: '{keyword}'
latex_cite_completion_display_line2: '{title}'

will produce a multi-line dialog with the keywords on the top line and the titles on the bottom
line. The keywords will be the searchable ones (as many people are used to searching and adding
citations via the citation BibTeX keys), and the titles can be used to check you have the right
item. Similarly, more complex formats can be set up (see comments in the code)