Adding Steady Forward Sync
This commit provides the possibility to perform a steady
forward sycning between the PDF file and the LaTeX sources
on mouse click. If the current source file is a LaTeX source
and the option steady_sync is on (by default = true) then it
will automatically open the PDF file.

In addition there is a new toggle command to switch the setting
on or off using

        ctrl+l, t, m(mouse)
        super+l, t, m(mouse)