Do not add parem after ref completion, unless asked to
Fixes a longtime annoyance of mine. If you type (\ref{, the ref completion
will by default add a ) after inserting the label. If you have paren
automatching on (which most people do), this is most likely not what you
want. But, you can set the "ref_add_parenthesis" to true if you do want it
(i.e. if you like the old behavior).