Add function to handle parsing %!TEX directives
This is a backend change in the way we handle %!TEX directives.

Code that was previously parsing %!TEX directives has been changed to
use the new function.

As part of this change, the engine and options used by the
traditional builder are not made available to all builders, though
only the traditional builder actually takes advantage of it.