Enable viewers to access the focus_st hack
This ensures that the hack only needs to be run if we know the viewer will steal the focus. `keep_focus`, set to `true` should mean "don't change focus as much as possible." I've also added a setting to disable the hack altogether, although this in no way guarantees that focus remains where you want it, since most PDF viewer will steal the focus on launch.