Added "newacronym" in glo_filter_commands
According to the Documentation of the glossaries latex package, \gls{}
command is used for both, the glossary entries and the acronyms.
Although the acronyms are handled by the \acr[full|short|long] command
in the plugin, these do not support that the first usage of the acronym
is written in full form and afterwards in short. Thus, \acr* commands
needs manual handling in latex, which is usually not desired.