Provide clearer math equation previews
This works by adding the TextAlphaBits and GraphicAlphaBits settings recommended in the Ghostscript documentation. In addition, if a new setting `"preview_math_hires"` is set to `true`, the image is rendered at 8x the requested density and then scaled down to the requested size. This seems to result in a clearer image. Finally, I've changed the default settings for the preview density and scale factor as image scaling in ST seems to add a small amount of distortion.

Over all, this should make the equation previews crisper on Windows and Linux (for some reason, they were fine for me on a Mac before this change).