Change the bracket completing algorithm to be less agressive
Per @r-stein's suggestion, the new algorithm defaults to only scanning 5 lines before the selection to determine if there is an unopened bracket that needs to be closed. This, hopefully, helps resolve the issues in #1011.

The number of lines scanned can be configured using the `smart_bracket_look_around` advanced setting, and the present behaviour of scanning the full document can be restored by using the `smart_bracket_scan_full_document` advanced setting.