Implements feature requested in #1043
Adds a command to wrap the selected text in a simple LaTeX command and automatically expands no selection to the current word. This also changes the keybindings from using snippets to using the command for emph, bold, etc. Additionally, I suppose, this makes new keybindings for new similar commands pretty trivial to implement. I've left the snippets in place so they can be used if desired, but obviously they do not have the same word-grabbing rules.