Created the parameter single_selection for latex_fill_all
command. It is a boolean indicating whether or not to ignore
multiple cursors/carets on the text. This is useful when creating
a snippet which defaults to creating multiple carets/cursors and
want to force the completion for some specific scope as `input`
or `cite`. Therefore we must to ignore the multi selection,
otherwise it will ignore the input scope specified.

Solves the issue:
1. https://github.com/SublimeText/LaTeXTools/issues/1161

For that solution, you just need to use this command on the
command palette, adding the `single_selection` argument:
```latex
[
	{
		"caption": "LaTeXTools: Insert Image",
		"command": "chain",
		"args": {
		    "commands": [
		        ["insert_snippet", {"contents": "\\begin{figure}\n\\centering\n\\includegraphics[scale=1.0]{$TM_SELECTED_TEXT$1}\n\\caption{$1}\n\\label{fig:$1}\n\\end{figure}"}],
		        ["latex_fill_all", {"completion_type": "input", "force": true, "single_selection" : true} ]
		    ]
		},
	},
]
```