Fix jump to PDF when compiling from non-.tex file
Previously when compiling from a non-.tex file (such as a .sty
file in which you are working), compiling would be succesful,
but trying to view it would give an error that it was impossible
to jump from something else than a .tex-file.
This is true when trying to jump with a keybinding, but in a new
compilation, this should just open the PDF at the same page
it already was (or the first page if the PDF wasn't open).
This change fixes that by checking whether the command was invoked
from a keybinding or not.