wrap selection in environment with newline stripped (#1666)
Resolves #1016

This commit automatically stripps a single newline from selection, when
wrapping it into an environment or command.

As a result outcome is the same regardless a line is selected with or without
trailing newline character - in other words: the caret is at the eol of
selected line or bol of line after.